requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
timm>=0.9.0
librosa>=0.10.0
soundfile>=0.12.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
PyYAML>=6.0
tqdm>=4.65.0
audioread>=3.0.0
